% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_plotly.R
\name{misc_plotly}
\alias{misc_plotly}
\title{Chart the Multiple Indicator Sigma Chart in Plotly}
\usage{
misc_plotly(
  data,
  hover_colour = "#00667B",
  control_colour = "#00667B",
  worse_colour = "#E46C0A",
  neutral_colour = "#A6A6A6",
  better_colour = "#00B0F0",
  y_dp = 2
)
}
\arguments{
\item{data}{A data.frame/data.table of the funnel data with the
columns indicator, group, numerator and denominator. See ?example_funnel_data
for an example and more information on the fields required.}

\item{hover_colour}{The background colour for the hoverinfo (default =
"#00667B")}

\item{control_colour}{The colour of the 95\% and 99\% control limits for the plot
(default = "#00667B")}

\item{worse_colour}{The colour of the bar in the unfavourable direction
(default = "#E46C0A")}

\item{neutral_colour}{The colour of the bar in the favourable direction
(default = "#A6A6A6")}

\item{better_colour}{The colour of the bar in the favourable direction
(default = "#00B0F0")}

\item{y_dp}{The number of decimal places displayed on the chart (default = 2)}
}
\value{
A MISC plotly output.
}
\description{
Chart the Multiple Indicator Sigma Chart in Plotly
}
\examples{
\dontrun{
  library(qiverse.qimisc)
  library(qiverse.data)
  misc_data <- misc_prep_data(example_funnel_data, example_indicator_data)
  misc_plotly(misc_data[group == "A"])
}
}
