% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_prep_data.R
\name{misc_prep_data}
\alias{misc_prep_data}
\title{Prepare data for the Multiple Indicator Sigma Chart}
\usage{
misc_prep_data(funnel_data, indicator_data)
}
\arguments{
\item{funnel_data}{A data.frame/data.table of the funnel data with the
columns indicator, group, numerator and denominator. See ?example_funnel_data
for an example and more information on the fields required.}

\item{indicator_data}{A data.frame/data.table of the indicator metadata with
the columns indicator_theme, indicator, multiplier, data_type and betteris.
See ?example_indicator_data for an example and more information on the fields
required.}
}
\value{
A data.table with the required fields for the MISC plotly chart.
}
\description{
Prepare data for the Multiple Indicator Sigma Chart
}
\examples{
\dontrun{
  library(qiverse.qimisc)
  library(qiverse.data)
  misc_data <- misc_prep_data(example_funnel_data, example_indicator_data)
}
}
